+++ backend/services/prompt_service.py
@@ -14,7 +14,7 @@ from database.agent_db import query_sub_agents, update_agent, \
     query_tools_by_ids
 from services.agent_service import get_enable_tool_id_by_agent_id
 from utils.config_utils import config_manager
-from database.utils import get_current_user_id
+from utils.auth_utils import get_current_user_id
 from fastapi import Header
