+++ backend/utils/auth_utils.py
@@ -2,6 +2,7 @@ import logging
 import os
 from typing import Optional
 
+from fastapi import Header
 import jwt
 
 # 获取Supabase配置
@@ -34,3 +35,18 @@ def get_current_user_id_from_token(authorization: str) -> Optional[str]:
     except Exception as e:
         logging.error(f"从令牌提取用户ID失败: {str(e)}")
         return None
+    
+def get_current_user_id(authorization: Optional[str] = None) -> tuple[str, str]:
+    DEFAULT_USER_ID = "user_id"
+    DEFAULT_TENANT_ID = "tenant_id"
+    
+    if authorization is None or authorization == Header(None):
+        return DEFAULT_USER_ID, DEFAULT_TENANT_ID
+        
+    try:
+        user_id = get_current_user_id_from_token(str(authorization))
+        tenant_id = user_id
+        return user_id, tenant_id
+    except Exception as e:
+        logging.error(f"获取用户ID失败: {str(e)}")
+        return DEFAULT_USER_ID, DEFAULT_TENANT_ID
